SOC-2378 | feat: changes after code review
