+++ src/test/java/com/wikia/webdriver/common/remote/operations/DeleteRemoteOperation.java
@@ -4,12 +4,13 @@ import com.wikia.webdriver.common.core.helpers.User;
 import org.apache.http.client.methods.HttpDelete;
 
 
-public class DeleteRemoteOperation extends BaseRemoteOperation {
+public class DeleteRemoteOperation extends BaseRemoteOperation implements RemoteOperation {
 
   DeleteRemoteOperation(User user) {
     super(user);
   }
 
+  @Override
   public String execute(final String url) {
     return super.execute(new HttpDelete(url));
   }
