+++ src/test/java/com/wikia/webdriver/common/remote/operations/LockPost.java
@@ -1,25 +1,10 @@
 package com.wikia.webdriver.common.remote.operations;
 
 import com.wikia.webdriver.common.core.helpers.User;
-import com.wikia.webdriver.common.remote.Discussions;
-import com.wikia.webdriver.common.remote.context.ThreadContext;
 
-public class LockPost {
-
-  static final String LOCK_POST_URL_SUFFIX = "%s/threads/%s/lock";
-
-  private final PutRemoteOperation remoteOperation;
+public class LockPost extends PostPadlock {
 
   public LockPost(User user) {
-    this.remoteOperation = new PutRemoteOperation(user);
-  }
-
-  public void execute(final ThreadContext context) {
-    final String url = buildUrl(context);
-    remoteOperation.execute(url);
-  }
-
-  private String buildUrl(final ThreadContext context) {
-    return Discussions.service(String.format(LOCK_POST_URL_SUFFIX, context.getSiteId(), context.getPostId()));
+    super(new PutRemoteOperation(user));
   }
 }
