+++ src/test/java/com/wikia/webdriver/common/remote/operations/RemoteOperation.java
@@ -1,0 +1,6 @@
+package com.wikia.webdriver.common.remote.operations;
+
+interface RemoteOperation {
+
+  String execute(final String url);
+}
