+++ src/test/java/com/wikia/webdriver/elements/mercury/components/discussions/common/DiscussionsConstants.java
@@ -2,6 +2,8 @@ package com.wikia.webdriver.elements.mercury.components.discussions.common;
 
 public final class DiscussionsConstants {
 
+  public static final String DESKTOP_RESOLUTION = "1920x1080";
+
   public static final long TIMEOUT = 10;
 
   private DiscussionsConstants() {
